+++ anonymous/anonymous/Form1.cs
@@ -41,86 +41,7 @@ namespace anonymous
         }
     }
 
-    interface IMatrix
-    {
-        IVector Multiply(IVector x);
-        IVector TMultiply(IVector x);
-        void test();
-    }
-
-    interface IVector
-    {
-        /*
-        static double operator *(IVector v1, IVector v2);
-        static IVector operator +(IVector v1, IVector v2);
-        static IVector operator *(IVector v1, double v2);
-        static IVector operator *(double v1, IVector v2);
-        */
-        double Norm();
-    }
-
-    public static class InputOutput
-    {
-        public static int formattype;
-        //Ввод плотной матрицы
-        public static void InputMatrix(int Flag, string FileName,out double[,] Plot) { Plot = new double[10,10]; }
-        //Ввод матрицы в профильном формате
-        public static void InputMatrix(int Flag, string FileName, out int[] ia, out double[] gg, out double[] diag)
-        {
-            gg = new double[10];
-            diag = new double[10];
-            ia = new int[10];
-        }
-        //Ввод матрицы в разреженном формате
-        public static void InputMatrix(int Flag, string FileName, out int[] ia, out int[] ja, out double[] gg, out double[] diag)
-        {
-            gg = new double[10];
-            diag = new double[10];
-            ia = new int[10];
-            ja = new int[10];
-        }
-        //Ввод матрицы в диагональном формате
-        //void InputMatrix(int Flag, string FileName, double[] gg, double[] diag);
-        //Ввод матрицы в ленточном формате
-        public static void InputMatrix(int Flag, string FileName, out double[] gg, out double[] diag)
-        {
-            gg = new double[10];
-            diag = new double[10];
-        }
-
-        //ввод вектора правой части
-        public static void InputRightPart(string FileName, out double[] pr)
-        {
-            pr = new double[10];
-        }
-    }
-
-    interface ISLAE
-    {
-        IMatrix Matrix { get; set; }
-        IVector RightPart { get; set; }
-    }
-
-    interface IIterationLogger
-    {
-        int IterationNumber { get; set; }
-        IVector Residual { get; set; }
-        IVector CurrentSolution { get; set; }
-        ISolver CurrentSolver { get; set; }
-    }
-
-    interface IPreconditioner : IMatrix
-    {
-        string Name { get; }
-        void Create(IMatrix matrix);
-    }
-
-    interface ISolver
-    {
-        IVector Solve(ISLAE slae, IVector initial, IIterationLogger logger, double eps, int maxiter);
-        string Name { get; }
-        IPreconditioner Preconditioner { get; set; }
-    }
+   
 
     class Matrix : IMatrix // Реализация интерфейса IMatrix
     {
